requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
playwright==1.40.0

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.0

# AI/ML
openai==1.3.0
anthropic==0.7.0

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.23

# Async support
asyncio
aiohttp==3.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1